feat: implements product management with CRUD operations

+++ database/migrations/00001_create_products_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    price BIGINT NOT NULL,
+    price INT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
